Sort out common util func/macro to bf_module_util
* Instead of exporting them from bf_switch_sysfs, now they are staticly
  linke by each module.
