+++ platform/barefoot/sonic-platform-modules-bfn-newport/modules/switch_interface/bf_sensor_driver.c
@@ -14,6 +14,7 @@
 #include <linux/slab.h>
 #include <linux/platform_device.h>
 #include "bf_switch_sysfs.h"
+#include "bf_module_util.h"
 #include "bf_sensor_driver.h"
 #include "bf_sensor_api.h"
